python/requirements.txt
# Core ML/AI libraries
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
ccxt>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# ONNX export
onnx>=1.14.0
onnxruntime>=1.15.0
